feat: impl rbac for background job

+++ src/infra/src/table/entity/background_job_partitions.rs
@@ -8,13 +8,13 @@ pub struct Model {
     #[sea_orm(primary_key, auto_increment = false)]
     pub job_id: String,
     #[sea_orm(primary_key, auto_increment = false)]
-    pub partition_id: i32,
+    pub partition_id: i64,
     pub start_time: i64,
     pub end_time: i64,
     pub created_at: i64,
     pub started_at: Option<i64>,
     pub ended_at: Option<i64>,
-    pub status: i32,
+    pub status: i64,
     pub result_path: Option<String>,
     #[sea_orm(column_type = "Text", nullable)]
     pub error_message: Option<String>,
